renamed dvgeo component
